Add full example to use the ORM
